:wrench: Utils can now find a tagtype without tag
What this means is that when given a string, it can iterate through all
that may apply on the server to get the right one.

Practically, it means that you can get information on something without
pinging it, which is especially useful with things like roles, where you
may not be able to ping it, or when pinging it may result in a lot of
anger towards you (*glances at @everyone*).