:package: Redone comments and styling
As said on the tin. There's some comments I've taken out, some I've
added in, and some parts which I've stylised.

The only significant changes that you would be able to find here are
packages of regex which are functionally identical to whatever was there
previously, with the advantage of being regex.

With this I've also disabled the LEGO and Tags extentions, due to lack
of use of the LEGO one, and the Tags one needs to be desperately
rewritten.